+++ packages/teamscale-javascript-instrumenter/src/vaccine/main.ts
@@ -1,10 +1,11 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore: DataWorker import
+// @ts-ignore: DataWorker import is handled by Esbuild---see `esbuild.mjs` and `workers.d.ts`
 import DataWorker from './worker/vaccine.worker.ts';
-import { IstanbulCoverageStore, makeProxy } from './Interceptor';
+import { makeProxy } from './Interceptor';
 import * as unload from 'unload';
 import { getWindow, universe, hasWindow, universeAttribute } from './utils';
 import { ProtocolMessageTypes } from './protocol';
+import { IstanbulCoverageStore } from './types';
 
 // Prepare our global JavaScript object. This will hold
 // a reference to the WebWorker thread.
@@ -25,6 +26,8 @@ function setWorker(worker: DataWorker): DataWorker {
 	return worker;
 }
 
+const interceptedStores: Set<string> = new Set<string>();
+
 /**
  * The function that intercepts changes to the Istanbul code coverage.
  * Also, the Web worker to forward the coverage information is started.
@@ -33,6 +36,14 @@ universe().makeCoverageInterceptor = function (coverage: IstanbulCoverageStore) 
 	// The `fileId` is used to map coverage and source maps. Note that
 	// a browser window (tab) can run multiple JavaScript files, with different source maps, ... .
 	const fileId = coverage.hash;
+
+	// Prevent adding the interceptor twice for the same file
+	if (interceptedStores.has(fileId)) {
+		console.log(`Coverage interceptor added twice for ${fileId}. This seems to be a bug in the instrumentation.`);
+		return;
+	} else {
+		interceptedStores.add(fileId);
+	}
 
 	if (!getWorker()) {
 		// Create the worker with the worker code
@@ -72,7 +83,10 @@ universe().makeCoverageInterceptor = function (coverage: IstanbulCoverageStore) 
 		})();
 	}
 
-	(function sendSourceMaps() {
+	(function sendSourceMapsAndCoverageObject() {
+		// Send the coverage object
+		getWorker().postMessage(`${ProtocolMessageTypes.ISTANBUL_COV_OBJECT} ${JSON.stringify(coverage)}`);
+
 		// Send the source maps
 		const sentMaps = universeAttribute('sentMaps', new Set());
 		if (coverage.inputSourceMap) {
@@ -105,5 +119,5 @@ universe().makeCoverageInterceptor = function (coverage: IstanbulCoverageStore) 
 		};
 	})();
 
-	return makeProxy(coverage, coverage, []);
+	return makeProxy(getWorker(), coverage.hash, coverage, []);
 };